Implement handling of HTTP probe code argument
This implements proper handling of the 'code' parameter that can be
provided to the HTTP probe action.

Previously, setting this didn't allow the probe to work when the
endpoint returned a 4xx or 5xx response code.

This adds an if statement to see whether the response code matches
the configured code property. If it does, it skips calling
resp.raise_for_status.
Signed-off-by: Nikolas Skoufis <nskoufis@seek.com.au>