Correctly handle non-error responses that don't match code
In the previous commit, I added code that handled when the response code
doesn't match the expected response code. However this doesn't work if
the actual response code is a non-4xx or 5xx code, which doesn't cause
raise_for_status to throw.

This removes the raise for status, and instead we just do a simple
comparison of the response code to the expected code, and log an error
if they don't match.

Signed-off-by: Nikolas Skoufis <nskoufis@seek.com.au>